Additional MechanismParams in HighLevelAPI

+++ src/Pkcs11Interop/LowLevelAPI/MechanismParams/CK_PKCS5_PBKD2_PARAMS.cs
@@ -69,6 +69,6 @@ namespace Net.Pkcs11Interop.LowLevelAPI.MechanismParams
         /// <summary>
         /// Length in bytes of the password information
         /// </summary>
-        public IntPtr PasswordLen;
+        public uint PasswordLen;
     }
 }
